application/transformation/requirements.txt
asyncio
typing
redis[hiredis]
python-dotenv
pydantic
PyPDF2
aiohttp
openai
groq
langchain
langchainhub
langsmith
pdfminer.six
pymupdf